test: organize step code in a reusable way
