+++ cypress/cypress/e2e/business-information-bc-registered-business.feature
@@ -1,9 +1,11 @@
-Feature: Business Information: Business name selection
+Feature: Business Information: BC registered business
 
   Background: List of business names is presented
     Given I navigate to the client application form
     When I select the option that says I have a BC registered business
     And I type in the first 3 characters of the business name
+    Then I am presented a list of business names filtered by the text typed in the business name field
+    And the button Next is disabled
 
   Scenario: Business in good standing
     When I select the name of a business in good standing from the filtered list
